html/semantics/embedded-content/the-video-element/posterloading/posterloading-lazy-scroll-into-view.html

<!DOCTYPE html>
<head>
  <title>Below viewport poster images with posterloading='lazy' load when scrolled into view</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<script>
  const t = async_test();

  let has_below_viewport_loaded = false;
  let has_window_loaded = false;

  window.addEventListener("load", t.step_func(function() {
    assert_false(has_window_loaded,
                 "The window load event should only fire once.");
    has_window_loaded = true;
    assert_false(has_below_viewport_loaded, 'Lazy poster image should not have loaded when the window load event fires.');
    document.getElementById('below_viewport').scrollIntoView();
  }));

  const below_viewport_img_onload = t.step_func_done(function() {
    assert_false(has_below_viewport_loaded,
               "The below_viewport element should load only once.");
    has_below_viewport_loaded = true;
  });
</script>

<body>
  <div style="height:10000px;"></div>
  <video id="below_viewport" poster="/images/blue.png?posterloading-lazy-below-viewport"
       posterloading="lazy"></video>
  <script>
    const ro = new ResizeObserver(entries => {
      for (let entry of entries) {
        const cr = entry.contentRect;
        if (cr.width === 300 && cr.height === 150)
          continue;
        else
          below_viewport_img_onload();
      }
    });

    ro.observe(document.getElementById('below_viewport'));
  </script>
</body>
